+++ components/Teaser/index.tsx
@@ -190,6 +190,8 @@ const Teaser = (props: TeaserProps) => {
           editable={editable}
           fallback_image_url={fallback_image_url}
           image_circle={layout?.image_circle}
+          radius_bottom={layout?.radius_bottom}
+          radius_top={layout?.radius_top}
           item={item}
           item_path={item_path}
           layout={layout}