+++ components/Teaser/types.ts
@@ -111,6 +111,8 @@ export type LayoutProps = {
   profile_image?: {
     enabled: boolean,
   };
+  radius_bottom?: boolean;
+  radius_top?: boolean;
   related?: Related;
   section?: Object;
   subject?: SubjectLayout;